+++ cmd/aws/batch/config.go
@@ -82,7 +82,7 @@ func DefaultConfig() Config {
 			},
 		},
 		JobRole: JobRoleConfig{
-			RoleName:           "FunnelEcsTaskRole",
+			RoleName:           "FunnelECSTaskRole",
 			DynamoDBPolicyName: "FunnelDynamoDB",
 			S3PolicyName:       "FunnelS3",
 		},
