+++ cmd/aws/batch/createall.go
@@ -55,7 +55,7 @@ var createCmd = &cobra.Command{
 		switch err.(type) {
 		case nil:
 			log.Info("Created ComputeEnvironment", "description", a)
-		case errResourceExists:
+		case awsutil.ErrResourceExists:
 			log.Error("ComputeEnvironment already exists", "description", a)
 		default:
 			return fmt.Errorf("failed to create ComputeEnvironment: %v", err)
@@ -65,7 +65,7 @@ var createCmd = &cobra.Command{
 		switch err.(type) {
 		case nil:
 			log.Info("Created JobQueue", "description", b)
-		case errResourceExists:
+		case awsutil.ErrResourceExists:
 			log.Error("JobQueue already exists", "description", b)
 		default:
 			return fmt.Errorf("failed to create JobQueue: %v", err)
@@ -75,7 +75,7 @@ var createCmd = &cobra.Command{
 		switch err.(type) {
 		case nil:
 			log.Info("Created JobDefinition", "description", c)
-		case errResourceExists:
+		case awsutil.ErrResourceExists:
 			log.Error("JobDefinition already exists", "description", c)
 		default:
 			return fmt.Errorf("failed to create JobDefinition: %v", err)
