+++ cmd/aws/batch/createjobdef.go
@@ -50,7 +50,7 @@ var jobdefCmd = &cobra.Command{
 		switch err.(type) {
 		case nil:
 			log.Info("Created JobDefinition", "description", c)
-		case errResourceExists:
+		case awsutil.ErrResourceExists:
 			log.Error("JobDefinition already exists", "description", c)
 		default:
 			return fmt.Errorf("failed to create JobDefinition: %v", err)
