+++ cmd/aws/beanstalk/deploy.go
@@ -16,16 +16,16 @@ var bucket = "strucka-dev"
 var confPath = ""
 
 func init() {
-	f := DeployCmd.Flags()
+	f := deployCmd.Flags()
 	f.StringVar(&image, "image", image, "Funnel docker image to deploy")
 	f.StringVar(&confPath, "config", confPath, "Config file")
 	f.StringVar(&bucket, "bucket", bucket, "S3 bucket in which to upload and store the Funnel source bundle")
 	f.StringVar(&region, "region", region, "AWS region in which to deploy Funnel server")
 }
 
-// DeployCmd represent the command responsible for deploying a Funnel server on
-// Amazon ElasticBeanstalk.
-var DeployCmd = &cobra.Command{
+// deployCmd represent the command responsible for deploying a Funnel server on
+// AWS Elastic Beanstalk.
+var deployCmd = &cobra.Command{
 	Use:   "deploy",
 	Short: "Deploy a Funnel server to Elastic Beanstalk.",
 	RunE: func(cmd *cobra.Command, args []string) error {
@@ -47,8 +47,8 @@ func deploy(ctx context.Context, image string, confPath string, bucket string, r
 	switch err.(type) {
 	case nil:
 		log.Info("Uploaded Application SourceBundle", "url", s3Zip)
-	case *awsutil.ErrResourceExists:
-		log.Info("Application SourceBundle already exists", "url", s3Zip)
+	case awsutil.ErrResourceExists:
+		log.Error("Application SourceBundle already exists", "url", s3Zip)
 	default:
 		return fmt.Errorf("failed to upload application source bundle: %v", err)
 	}
@@ -57,7 +57,7 @@ func deploy(ctx context.Context, image string, confPath string, bucket string, r
 	switch err.(type) {
 	case nil:
 		log.Info("Created Application", "description", ar)
-	case *awsutil.ErrResourceExists:
+	case awsutil.ErrResourceExists:
 		log.Error("Application already exists", "description", ar)
 	default:
 		return fmt.Errorf("error creating application: %v", err)
@@ -67,7 +67,7 @@ func deploy(ctx context.Context, image string, confPath string, bucket string, r
 	switch err.(type) {
 	case nil:
 		log.Info("Created Application Environment", "description", er)
-	case *awsutil.ErrResourceExists:
+	case awsutil.ErrResourceExists:
 		log.Error("Application Environment already exists", "description", er)
 	default:
 		return fmt.Errorf("error creating environment: %v", err)
